updates the eos_vxlan module
This commit updates the eos_vxlan module to use the new pyeapi client for
eapi.  The new eos_vxlan module is now a super set of the base interface
in pyeapi which expose two new attributes: enable and description.  This
means that the eos_interface module is no longer required to manage the
base interface attributes of the vxlan interface