updates the eos_interface module
The commit refacters the eos_interface module to be a generic interface
module that can be used with any EOS interface.  The parameters that were
part of the original interface module are being refactored into the
eos_ethernet module.

This commit can cause distruption with existing modules due to the
refractoring of the parameters from interface to ethernet

This commit also adds an initial set of test cases for the
eos_interface module