updates facts documentation and adds arguments
This commit adds two arguments to eos_facts - include and exclude. The include
argument will specify the list of facts to include in the module run.  The
exclude argument will specify the list of facts to exclude in the module
run.  The two arguments are mutually exclusive.  Each argument can accept
a comma delimited list of facts to include or exclude