adds capability to configure interface description to empty string
This commit addresses setting the interface description to an empty string.
Prior to this patch, setting the description argument to '' would not configure
the interface description properly.  After this patch is applied, the
interface description can now be configured description='' to remove the
configured string.

This commit applies to all modules that support an interface description
which includes:

    * eos_interface
    * eos_vxlan
    * eos_ethernet
    * eos_portchannel

This commit also updates the module test cases to test for configuring the
interface descrption with an empty string